fix: fix some more errors
